Time complexity experiment
This was an attempt at reducing time complexity of the WFLW Neural Net training program by splitting the training data. It was not particularly successful and it was later discovered that this type of approach is generally used more for LLMs and larger models with multiple servers to run on.